requirements.txt
cvxpy==1.1.15
matplotlib==3.3.3
numpy==1.19.2
pandas==1.4.1
prettytable==0.9.2
PyYAML==6.0
scikit_learn==0.23.2
torch==1.7.0
tqdm==4.62.3
